split service core for all services

+++ services/system-service/src/main/java/org/shay/education/system/service/impl/ApplicationServiceImpl.java
@@ -37,6 +37,6 @@ public class ApplicationServiceImpl extends ServiceImpl<ApplicationMapper, Appli
             ApplicationDTO dto = model.toDTO();
             dtoList.add(dto);
         }
-        return new PagedDTO<>((int) paged.getTotal(), dtoList);
+        return new PagedDTO<>(paged.getTotal(), dtoList);
     }
 }
